Update Red Pikmin
Make it so idle animation has leaf hanging down. Not sure
if I like this implementation, but it makes the left/right frames look
better which I also added. For the left/right frame, I had to shift over
the drawing, which I'm not sure how it'll work out when drawing the pixels
Also exported a gif